\name{FactoInvestigate-package}
\alias{FactoInvestigate-package}
\alias{FactoInvestigate}
\docType{package}
\title{
\packageTitle{FactoInvestigate}
}
\description{
\packageDescription{FactoInvestigate}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FactoInvestigate}
\packageIndices{FactoInvestigate}
An overview of how to use the package, including the most important functions
}
\author{
\packageAuthor{FactoInvestigate}

Maintainer: \packageMaintainer{FactoInvestigate}
}
\keyword{package}

\seealso{
\code{\link[FactoMineR:FactoMineR-package]{FactoMineR}}
}